feat(dev): install packages usefull for development with docker
Install avahi and dropbear for connecting through ssh to the device
using the hostname. Install docker and docker-compose for the container
support.

There is the need to increase the disk size of the image for the new
packages, but since this is for a development image we do not specify it
in the configuration yet, and can be set in the menu config afterwards
to the required size. Using 512 MB is more than enough.